simplify change to only addong ConsumerTrace
