--- Src/zipkin4net/Src/BaseStandardTrace.cs
@@ -1,20 +1,0 @@
-﻿using zipkin4net.Annotation;
-using System;
-
-namespace zipkin4net
-{
-    public class BaseStandardTrace
-    {
-        public virtual Trace Trace { internal set;  get; }
-
-        public void AddAnnotation(IAnnotation annotation)
-        {
-            Trace.Record(annotation);
-        }
-
-        public virtual void Error(Exception ex)
-        {
-            Trace.Record(Annotations.Tag("error", ex.Message));
-        }
-    }
-}
