+++ Src/zipkin4net/Src/ServerTrace.cs
@@ -1,11 +1,12 @@
-﻿using System;
+﻿using zipkin4net.Annotation;
+using System;
 using System.Threading.Tasks;
 
 namespace zipkin4net
 {
-    public class ServerTrace : BaseStandardTrace, IDisposable
+    public class ServerTrace : IDisposable
     {
-        public override Trace Trace
+        public Trace Trace
         {
             get
             {
@@ -20,9 +21,19 @@ namespace zipkin4net
             Trace.Record(Annotations.Rpc(rpc));
         }
 
+        public void AddAnnotation(IAnnotation annotation)
+        {
+            Trace.Record(annotation);
+        }
+
         public void Dispose()
         {
             Trace.Record(Annotations.ServerSend());
+        }
+
+        public virtual void Error(Exception ex)
+        {
+            Trace.RecordAnnotation(Annotations.Tag("error", ex.Message));
         }
     }
